requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.2.2
django-enumchoicefield==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
drf-yasg==1.21.7
filelock==3.12.2
flake8==6.0.0
identify==2.5.24
idna==3.4
inflection==0.5.1
install==1.3.5
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mccabe==0.7.0
nodeenv==1.8.0
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
platformdirs==3.7.0
pre-commit==3.3.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.23.1
